requirements.txt
streamlit
openai
langchain
langchain-openai
langchain-community
langchain-core
langchain-text-splitters
faiss-cpu
python-dotenv
requests
beautifulsoup4
lxml
